+++ compile/program.c
@@ -63,7 +63,7 @@ main_func_t compile_file(gcc_ctx_t *ctx, jmp_buf *on_err, sss_file_t *f, ast_t *
     gcc_return(main_block, NULL, gcc_zero(ctx, gcc_type(ctx, INT)));
 
     // Actually compile the functions:
-    for (uint32_t i = 1; i <= Table_length(&env->global->ast_functions); i++) {
+    for (int64_t i = 1; i <= Table_length(&env->global->ast_functions); i++) {
         struct {const char *key; func_context_t *value;} *entry = Table_str_entry(&env->global->ast_functions, i);
         ast_t *ast = (ast_t*)strtol(entry->key, NULL, 16);
         compile_function(&entry->value->env, entry->value->func, ast);
@@ -118,7 +118,7 @@ void compile_object_file(gcc_ctx_t *ctx, jmp_buf *on_err, sss_file_t *f, ast_t *
     gcc_return(finished_loading_block, NULL, gcc_rval(module_var));
 
     // Actually compile the functions:
-    for (uint32_t i = 1; i <= Table_length(&env->global->ast_functions); i++) {
+    for (int64_t i = 1; i <= Table_length(&env->global->ast_functions); i++) {
         struct {const char*key; func_context_t *value;} *entry = Table_str_entry(&env->global->ast_functions, i);
         ast_t *ast = (ast_t*)strtol(entry->key, NULL, 16);
         compile_function(&entry->value->env, entry->value->func, ast);
