+++ environment.c
@@ -351,7 +351,7 @@ env_t *scope_with_type(env_t *env, sss_type_t *t)
         if (LENGTH(members) > 0 && !Table_str_get(ns, ith(members, 0).name))
             populate_tagged_union_constructors(env, t);
     }
-    for (uint32_t i = 1; i <= Table_length(ns); i++) {
+    for (int64_t i = 1; i <= Table_length(ns); i++) {
         struct { const char *key; binding_t *value; } *entry = Table_entry(ns, i);
         if (!Table_str_get(fresh->bindings, entry->key))
             Table_str_set(fresh->bindings, entry->key, entry->value);
@@ -362,7 +362,7 @@ env_t *scope_with_type(env_t *env, sss_type_t *t)
 static void copy_global_bindings(table_t *dest, table_t *src)
 {
     for (; src; src = src->fallback) {
-        for (uint32_t i = 1; i <= Table_length(src); i++) {
+        for (int64_t i = 1; i <= Table_length(src); i++) {
             struct { const char *key; binding_t *value;} *entry = Table_str_entry(src, i);
             if (entry->value->visible_in_closures)
                 Table_str_set(dest, entry->key, entry->value);
@@ -650,7 +650,7 @@ const char *spellcheck(table_t *ns, const char *word)
     const char *closest = NULL;
     int closest_dist = 0;
     for (; ns; ns = ns->fallback) {
-        for (uint32_t i = 1; i <= Table_length(ns); i++) {
+        for (int64_t i = 1; i <= Table_length(ns); i++) {
             struct { const char *key; binding_t *value;} *entry = Table_str_entry(ns, i);
             int dist = word_distance(word, entry->key);
             if (closest == NULL || dist < closest_dist) {
